refactor(data): use abstract class for paging source
to remove re-implementation of the same code
