+++ app/src/main/java/io/github/kirasok/photogallerycompose/feature_viewer/data/repository/paging/SearchPagingSource.kt
@@ -1,35 +1,13 @@
 package io.github.kirasok.photogallerycompose.feature_viewer.data.repository.paging
 
-import android.util.Log
-import androidx.paging.PagingSource
-import androidx.paging.PagingState
+import io.github.kirasok.photogallerycompose.feature_viewer.data.dto.FlickrResponse
 import io.github.kirasok.photogallerycompose.feature_viewer.data.source.PhotoRemoteDataSource
-import io.github.kirasok.photogallerycompose.feature_viewer.domain.model.Photo
-import retrofit2.HttpException
-import java.io.IOException
 
 class SearchPagingSource(
   private val remoteDataSource: PhotoRemoteDataSource,
   private val query: String,
 ) :
-  PagingSource<Int, Photo>() {
-  override fun getRefreshKey(state: PagingState<Int, Photo>): Int? = state.anchorPosition
-
-  override suspend fun load(params: LoadParams<Int>): LoadResult<Int, Photo> = try {
-    val currentPage = params.key ?: 1
-    val photos = remoteDataSource.searchPhotos(query, currentPage, params.loadSize)
-    Log.i(null, "Got photos")
-    Log.d(null, photos.toString())
-    LoadResult.Page(
-      data = photos.photosResponse.photos,
-      prevKey = if (currentPage == 1) null else currentPage - 1,
-      nextKey = if (currentPage != photos.photosResponse.pages) currentPage + 1 else null
-    )
-  } catch (exception: IOException) {
-    Log.e(null, exception.toString())
-    LoadResult.Error(exception)
-  } catch (exception: HttpException) {
-    Log.e(null, exception.toString())
-    LoadResult.Error(exception)
-  }
+  PhotoPagingSource() {
+  override suspend fun request(loadSize: Int, currentPage: Int): FlickrResponse =
+    remoteDataSource.searchPhotos(query, currentPage, loadSize)
 }
